Fix zero metrics (#5)

+++ src/datasets/bert_dataset.py
@@ -42,6 +42,6 @@ class BERTDataset(Dataset):
             padding=True,
             return_tensors="pt",
         )
-        tokenized_batch["labels"] = torch.tensor(labels)
+        tokenized_batch["labels"] = torch.tensor(labels, dtype=torch.float)
 
         return tokenized_batch
